add RequestValidator, still need to upgrade

+++ backend/src/main/java/com/lerhyd/dngame/controller/EntryController.java
@@ -4,6 +4,7 @@ import com.lerhyd.dngame.dao.*;
 import com.lerhyd.dngame.info.EntryInfo;
 import com.lerhyd.dngame.model.*;
 import com.lerhyd.dngame.request.EntryReq;
+import com.lerhyd.dngame.validator.RequestValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -41,32 +42,36 @@ public class EntryController {
     }
 
     @PostMapping("/entry/add")
-    public void addEntry(@RequestBody EntryReq entryReq)
+    public boolean addEntry(@RequestBody EntryReq entryReq)
     {
-        //check if Entry's Person exists
-        boolean isExists = personDao.existsByNameAndSurnameAndPatronymicAndSex(entryReq.getVictimName(),
-                                                                                entryReq.getVictimSername(),
-                                                                                entryReq.getVictimPatr(),
-                                                                                entryReq.isVictimSex());
-        if (isExists){
-            Entry entry = getFormedEntry(entryReq, false);
-            entryDao.save(entry);
-            newsDao.save(generateNewsFromEntry(entry));
-            boolean isCriminal = personDao.checkPersonIfCriminal(entryReq.getVictimName(),
-                                            entryReq.getVictimSername(),
-                                            entryReq.getVictimPatr(),
-                                            entryReq.isVictimSex());
-            if (isCriminal){
-                kiraDao.deletePoints(10, entryReq.getKiraId());
-            } else {
-                kiraDao.addPoints(10, entryReq.getKiraId());
-            }
+        if (RequestValidator.validateEntryReq(entryReq, entryDao, kiraDao)) {
 
-        } else {
-            Entry entry = getFormedEntry(entryReq, true);
-            entryDao.save(entry);
-            kiraDao.deletePoints(30, entryReq.getKiraId());
+            //check if Entry's Person exists
+            boolean isExists = personDao.existsByNameAndSurnameAndPatronymicAndSex(entryReq.getVictimName(),
+                    entryReq.getVictimSername(),
+                    entryReq.getVictimPatr(),
+                    entryReq.isVictimSex());
+            if (isExists) {
+                Entry entry = getFormedEntry(entryReq, false);
+                entryDao.save(entry);
+                newsDao.save(generateNewsFromEntry(entry));
+                boolean isCriminal = personDao.checkPersonIfCriminal(entryReq.getVictimName(),
+                        entryReq.getVictimSername(),
+                        entryReq.getVictimPatr(),
+                        entryReq.isVictimSex());
+                if (isCriminal) {
+                    kiraDao.deletePoints(10, entryReq.getKiraId());
+                } else {
+                    kiraDao.addPoints(10, entryReq.getKiraId());
+                }
+
+            } else {
+                Entry entry = getFormedEntry(entryReq, true);
+                entryDao.save(entry);
+                kiraDao.deletePoints(30, entryReq.getKiraId());
+            }
         }
+        return false;
     }
 
     private boolean checkIfLose(long kiraId){
